requirements.txt
streamlit==1.51.0
python-dotenv==1.2.1
pypdf==6.4.0
sentence-transformers==2.2.2
faiss-cpu==1.13.0
transformers==4.37.1
requests>=2.28.0